Crate rusoto_importexport

AWS Import/Export Service AWS Import/Export accelerates transferring large amounts of data between the AWS cloud and portable storage devices that you mail to us. AWS Import/Export transfers data directly onto and off of your storage devices using Amazon's high-speed internal network and bypassing the Internet. For large data sets, AWS Import/Export is often faster than Internet transfer and more cost effective than upgrading your connectivity.

If you're using the service, you're probably looking for ImportExportClient and ImportExport.

Structs

Artifact
A discrete item that contains the description and URL of an artifact (such as a PDF).

CancelJobInput
Input structure for the CancelJob operation.

CancelJobOutput
Output structure for the CancelJob operation.

CreateJobInput
Input structure for the CreateJob operation.

CreateJobOutput
Output structure for the CreateJob operation.

GetShippingLabelInput

GetShippingLabelOutput

GetStatusInput
Input structure for the GetStatus operation.

GetStatusOutput
Output structure for the GetStatus operation.

ImportExportClient
A client for the AWS Import/Export API.

Job
Representation of a job returned by the ListJobs operation.

ListJobsInput
Input structure for the ListJobs operation.

ListJobsOutput
Output structure for the ListJobs operation.

UpdateJobInput
Input structure for the UpateJob operation.

UpdateJobOutput
Output structure for the UpateJob operation.

Enums

CancelJobError
Errors returned by CancelJob

CreateJobError
Errors returned by CreateJob

GetShippingLabelError
Errors returned by GetShippingLabel

GetStatusError
Errors returned by GetStatus

ListJobsError
Errors returned by ListJobs

UpdateJobError
Errors returned by UpdateJob

Traits

ImportExport
Trait representing the capabilities of the AWS Import/Export API. AWS Import/Export clients implement this trait.