Enum rusoto_importexport::CreateJobError

pub enum CreateJobError {
    BucketPermission(String),
    CreateJobQuotaExceeded(String),
    InvalidAccessKeyId(String),
    InvalidAddress(String),
    InvalidCustoms(String),
    InvalidFileSystem(String),
    InvalidJobId(String),
    InvalidManifestField(String),
    InvalidParameter(String),
    InvalidVersion(String),
    MalformedManifest(String),
    MissingCustoms(String),
    MissingManifestField(String),
    MissingParameter(String),
    MultipleRegions(String),
    NoSuchBucket(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Errors returned by CreateJob

Variants
BucketPermission(String)

The account specified does not have the appropriate bucket permissions.

CreateJobQuotaExceeded(String)

Each account can create only a certain number of jobs per day. If you need to create more than this, please contact awsimportexport@amazon.com to explain your particular use case.

InvalidAccessKeyId(String)

The AWS Access Key ID specified in the request did not match the manifest's accessKeyId value. The manifest and the request authentication must use the same AWS Access Key ID.

InvalidAddress(String)

The address specified in the manifest is invalid.

InvalidCustoms(String)

One or more customs parameters was invalid. Please correct and resubmit.

InvalidFileSystem(String)

File system specified in export manifest is invalid.

InvalidJobId(String)

The JOBID was missing, not found, or not associated with the AWS account.

InvalidManifestField(String)

One or more manifest fields was invalid. Please correct and resubmit.

InvalidParameter(String)

One or more parameters had an invalid value.

InvalidVersion(String)

The client tool version is invalid.

MalformedManifest(String)

Your manifest is not well-formed.

MissingCustoms(String)

One or more required customs parameters was missing from the manifest.

MissingManifestField(String)

One or more required fields were missing from the manifest file. Please correct and resubmit.

MissingParameter(String)

One or more required parameters was missing from the request.

MultipleRegions(String)

Your manifest file contained buckets from multiple regions. A job is restricted to buckets from one region. Please correct and resubmit.

NoSuchBucket(String)

The specified bucket does not exist. Create the specified bucket or change the manifest's bucket, exportBucket, or logBucket field to a bucket that the account, as specified by the manifest's Access Key ID, has write permissions to.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl CreateJobError

pub fn from_body(body: &str) -> CreateJobError