Enum rusoto_importexport::GetStatusError

pub enum GetStatusError {
    ExpiredJobId(String),
    InvalidAccessKeyId(String),
    InvalidJobId(String),
    InvalidVersion(String),
    CanceledJobId(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Errors returned by GetStatus

Variants
ExpiredJobId(String)

Indicates that the specified job has expired out of the system.

InvalidAccessKeyId(String)

The AWS Access Key ID specified in the request did not match the manifest's accessKeyId value. The manifest and the request authentication must use the same AWS Access Key ID.

InvalidJobId(String)

The JOBID was missing, not found, or not associated with the AWS account.

InvalidVersion(String)

The client tool version is invalid.

CanceledJobId(String)

The specified job ID has been canceled and is no longer valid.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl GetStatusError

fn from_body(body: &str) -> GetStatusError