Enum rusoto_importexport::UpdateJobError

pub enum UpdateJobError {
    UnableToUpdateJobId(String),
    InvalidFileSystem(String),
    ExpiredJobId(String),
    InvalidCustoms(String),
    InvalidManifestField(String),
    InvalidParameter(String),
    MissingCustoms(String),
    MissingManifestField(String),
    MissingParameter(String),
    InvalidAccessKeyId(String),
    InvalidJobId(String),
    BucketPermission(String),
    InvalidAddress(String),
    InvalidVersion(String),
    NoSuchBucket(String),
    CanceledJobId(String),
    MultipleRegions(String),
    MalformedManifest(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Errors returned by UpdateJob

Variants
UnableToUpdateJobId(String)

AWS Import/Export cannot update the job

InvalidFileSystem(String)

File system specified in export manifest is invalid.

ExpiredJobId(String)

Indicates that the specified job has expired out of the system.

InvalidCustoms(String)

One or more customs parameters was invalid. Please correct and resubmit.

InvalidManifestField(String)

One or more manifest fields was invalid. Please correct and resubmit.

InvalidParameter(String)

One or more parameters had an invalid value.

MissingCustoms(String)

One or more required customs parameters was missing from the manifest.

MissingManifestField(String)

One or more required fields were missing from the manifest file. Please correct and resubmit.

MissingParameter(String)

One or more required parameters was missing from the request.

InvalidAccessKeyId(String)

The AWS Access Key ID specified in the request did not match the manifest's accessKeyId value. The manifest and the request authentication must use the same AWS Access Key ID.

InvalidJobId(String)

The JOBID was missing, not found, or not associated with the AWS account.

BucketPermission(String)

The account specified does not have the appropriate bucket permissions.

InvalidAddress(String)

The address specified in the manifest is invalid.

InvalidVersion(String)

The client tool version is invalid.

NoSuchBucket(String)

The specified bucket does not exist. Create the specified bucket or change the manifest's bucket, exportBucket, or logBucket field to a bucket that the account, as specified by the manifest's Access Key ID, has write permissions to.

CanceledJobId(String)

The specified job ID has been canceled and is no longer valid.

MultipleRegions(String)

Your manifest file contained buckets from multiple regions. A job is restricted to buckets from one region. Please correct and resubmit.

MalformedManifest(String)

Your manifest is not well-formed.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl UpdateJobError

fn from_body(body: &str) -> UpdateJobError