Struct rusoto_importexport::ImportExportClient

pub struct ImportExportClient { /* fields omitted */ }

A client for the AWS Import/Export API.

Methods

impl ImportExportClient

pub fn new(region: Region) -> ImportExportClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> ImportExportClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,